test(repo): update e2e tests for new architecture
