+++ e2e-tests/fixtures/samples/thymian.config.yaml
@@ -1,13 +1,15 @@
+specifications:
+  - type: openapi
+    location: test.openapi.yaml
+ruleSets:
+  - '@thymian/rfc-9110-rules'
 plugins:
   '@thymian/http-linter': {}
-  '@thymian/openapi':
-    options:
-      descriptions:
-        - source: test.openapi.yaml
+  '@thymian/openapi': {}
   '@thymian/request-dispatcher': {}
   '@thymian/sampler': {}
+  '@thymian/http-tester': {}
   '@thymian/reporter':
     options:
       formatters:
         text: {}
-  '@thymian/format-validator': {}
