+++ e2e-tests/fixtures/static-lint/thymian.config.yaml
@@ -1,9 +1,11 @@
+specifications:
+  - type: openapi
+    location: test.openapi.yaml
+ruleSets:
+  - '@thymian/rfc-9110-rules'
 plugins:
   '@thymian/http-linter': {}
-  '@thymian/openapi':
-    options:
-      descriptions:
-        - source: test.openapi.yaml
+  '@thymian/openapi': {}
   '@thymian/reporter':
     options:
       formatters:
